requirements.txt
yfinance==0.1.54

python-dateutil~=2.8.1
setuptools~=54.0.0
ruamel.yaml
qrcode~=6.1
Pillow
pyautogui
requests
numpy
stomp.py
selenium
xlrd